BERT/requirements.txt
ml_collections
tqdm
tensorboard
scipy
pandas
numpy
matplotlib
certifi==2021.5.30
cycler==0.10.0
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
scikit-learn==1.0.1
six==1.16.0
